+++ configs/vae_config.py
@@ -1,0 +1,7 @@
+dataset_path = "~/datasets"
+batch_size = 100
+x_dim = 784
+hidden_dim = 400
+latent_dim = 200
+lr = 1e-3
+epochs = 30
