+++ models/vae.py
@@ -1,0 +1,66 @@
+import torch
+import torch.nn as nn
+
+
+class Encoder(nn.Module):
+
+    def __init__(self, input_dim: int, hidden_dim: int, latent_dim: int):
+        super(Encoder, self).__init__()
+
+        self.FC_input = nn.Linear(input_dim, hidden_dim)
+        self.FC_input2 = nn.Linear(hidden_dim, hidden_dim)
+        self.FC_mean = nn.Linear(hidden_dim, latent_dim)
+        self.FC_var = nn.Linear(hidden_dim, latent_dim)
+
+        self.LeakyReLU = nn.LeakyReLU(0.2)
+
+        self.training = True
+
+    def forward(self, x: torch.Tensor) -> tuple[torch.Tensor, torch.Tensor]:
+        h_ = self.LeakyReLU(self.FC_input(x))
+        h_ = self.LeakyReLU(self.FC_input2(h_))
+        mean = self.FC_mean(h_)
+        log_var = self.FC_var(h_)  # encoder produces mean and log of variance
+        #             (i.e., parateters of simple tractable normal distribution "q"
+
+        return mean, log_var
+
+
+class Decoder(nn.Module):
+    def __init__(self, latent_dim: int, hidden_dim: int, output_dim: int):
+        super(Decoder, self).__init__()
+        self.FC_hidden = nn.Linear(latent_dim, hidden_dim)
+        self.FC_hidden2 = nn.Linear(hidden_dim, hidden_dim)
+        self.FC_output = nn.Linear(hidden_dim, output_dim)
+
+        self.LeakyReLU = nn.LeakyReLU(0.2)
+
+    def forward(self, x: torch.Tensor) -> torch.Tensor:
+        h = self.LeakyReLU(self.FC_hidden(x))
+        h = self.LeakyReLU(self.FC_hidden2(h))
+
+        x_hat = torch.sigmoid(self.FC_output(h))
+        return x_hat
+
+
+class Model(nn.Module):
+
+    def __init__(self, encoder: nn.Module, decoder: nn.Module, device: str):
+        super(Model, self).__init__()
+        self._device = device
+        self.encoder = encoder
+        self.decoder = decoder
+
+    def reparameterization(self, mean, var):
+        epsilon = torch.randn_like(var).to(self._device)  # sampling epsilon
+        z = mean + var * epsilon  # reparameterization trick
+        return z
+
+    def forward(self, x: torch.Tensor) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
+        mean, log_var = self.encoder(x)
+        z = self.reparameterization(
+            mean, torch.exp(0.5 * log_var)
+        )  # takes exponential function (log var -> var)
+        x_hat = self.decoder(z)
+
+        return x_hat, mean, log_var
