--- vae.py
@@ -1,153 +1,0 @@
-import numpy as np
-import torch
-import torch.nn as nn
-import torchvision.transforms as transforms
-from torch.utils.data import DataLoader
-from torchvision.datasets import MNIST
-from torchvision.utils import make_grid, save_image
-from tqdm import tqdm
-
-from torch.optim import Adam
-
-# Model Hyperparameters
-
-dataset_path = "~/datasets"
-cuda = True
-_device = torch.device("cuda" if cuda else "cpu")
-
-batch_size = 100
-x_dim = 784
-hidden_dim = 400
-latent_dim = 200
-
-lr = 1e-3
-epochs = 30
-
-mnist_transform = transforms.Compose(
-    [
-        transforms.ToTensor(),
-    ]
-)
-
-kwargs = {"num_workers": 1, "pin_memory": True}
-
-train_dataset = MNIST(
-    dataset_path, transform=mnist_transform, train=True, download=True
-)
-test_dataset = MNIST(
-    dataset_path, transform=mnist_transform, train=False, download=True
-)
-
-train_loader = DataLoader(
-    dataset=train_dataset, batch_size=batch_size, shuffle=True, **kwargs
-)
-test_loader = DataLoader(
-    dataset=test_dataset, batch_size=batch_size, shuffle=False, **kwargs
-)
-
-
-class Encoder(nn.Module):
-
-    def __init__(self, input_dim, hidden_dim, latent_dim):
-        super(Encoder, self).__init__()
-
-        self.FC_input = nn.Linear(input_dim, hidden_dim)
-        self.FC_input2 = nn.Linear(hidden_dim, hidden_dim)
-        self.FC_mean = nn.Linear(hidden_dim, latent_dim)
-        self.FC_var = nn.Linear(hidden_dim, latent_dim)
-
-        self.LeakyReLU = nn.LeakyReLU(0.2)
-
-        self.training = True
-
-    def forward(self, x):
-        h_ = self.LeakyReLU(self.FC_input(x))
-        h_ = self.LeakyReLU(self.FC_input2(h_))
-        mean = self.FC_mean(h_)
-        log_var = self.FC_var(h_)  # encoder produces mean and log of variance
-        #             (i.e., parateters of simple tractable normal distribution "q"
-
-        return mean, log_var
-
-
-class Decoder(nn.Module):
-    def __init__(self, latent_dim, hidden_dim, output_dim):
-        super(Decoder, self).__init__()
-        self.FC_hidden = nn.Linear(latent_dim, hidden_dim)
-        self.FC_hidden2 = nn.Linear(hidden_dim, hidden_dim)
-        self.FC_output = nn.Linear(hidden_dim, output_dim)
-
-        self.LeakyReLU = nn.LeakyReLU(0.2)
-
-    def forward(self, x):
-        h = self.LeakyReLU(self.FC_hidden(x))
-        h = self.LeakyReLU(self.FC_hidden2(h))
-
-        x_hat = torch.sigmoid(self.FC_output(h))
-        return x_hat
-
-
-class Model(nn.Module):
-    def __init__(self, Encoder, Decoder):
-        super(Model, self).__init__()
-        self.Encoder = Encoder
-        self.Decoder = Decoder
-
-    def reparameterization(self, mean, var):
-        epsilon = torch.randn_like(var).to(_device)  # sampling epsilon
-        z = mean + var * epsilon  # reparameterization trick
-        return z
-
-    def forward(self, x):
-        mean, log_var = self.Encoder(x)
-        z = self.reparameterization(
-            mean, torch.exp(0.5 * log_var)
-        )  # takes exponential function (log var -> var)
-        x_hat = self.Decoder(z)
-
-        return x_hat, mean, log_var
-
-encoder = Encoder(input_dim=x_dim, hidden_dim=hidden_dim, latent_dim=latent_dim)
-decoder = Decoder(latent_dim=latent_dim, hidden_dim=hidden_dim, output_dim=x_dim)
-model = Model(Encoder=encoder, Decoder=decoder).to(_device)
-
-
-BCE_loss = nn.BCELoss()
-
-
-def loss_function(x, x_hat, mean, log_var):
-    reproduction_loss = nn.functional.binary_cross_entropy(x_hat, x, reduction="sum")
-    KLD = -0.5 * torch.sum(1 + log_var - mean.pow(2) - log_var.exp())
-
-    return reproduction_loss + KLD
-
-
-optimizer = Adam(model.parameters(), lr=lr)
-print("Start training VAE...")
-model.train()
-
-for epoch in range(epochs):
-    overall_loss = 0
-    for batch_idx, (x, _) in enumerate(train_loader):
-        x = x.view(batch_size, x_dim)
-        x = x.to(_device)
-
-        optimizer.zero_grad()
-
-        x_hat, mean, log_var = model(x)
-        loss = loss_function(x, x_hat, mean, log_var)
-
-        overall_loss += loss.item()
-
-        loss.backward()
-        optimizer.step()
-
-    print(
-        "\tEpoch",
-        epoch + 1,
-        "complete!",
-        "\tAverage Loss: ",
-        overall_loss / (batch_idx * batch_size),
-    )
-
-print("Finish!!")
